Fix name of overridden method.
In pycparser's commit 4a6afa0587bfb340dad5838f847e71704ea48ea7, the
method p_direct_declarator_5 was renamed to p_direct_declarator_6.
Because pycparserext didn't track this rename, sometimes the method from
the base class, which was supposed to be replaced, was called.  Also,
this broke the rarely-used "int A[*]" VLA syntax, because the method for
that from the base class was replaced instead.