Improve support for __extension__ keyword
This keyword is allowed to appear before expressions in GNU C and
serves as a hint to the compiler to not issue a warning in
case a GNU C specific extension is used, cf.:
https://gcc.gnu.org/onlinedocs/gcc/Alternate-Keywords.html

The two additional parser rules in this commit will allow to parse
that keyword correctly when it appears before an expression.

For now, the keyword is not reproduced when code is generated from
the AST, but this should not be a problem as it is just controlling
compiler warnings and does not affect semantics (as outlined above).